rename chapter8 to chapter7
