+++ src/main/webapp/chapter7/index.zul
@@ -1,0 +1,9 @@
+<?link rel="stylesheet" type="text/css" href="/style.css"?>
+<!-- protect page by the authentication init  -->
+<?init class="org.zkoss.essentials.chapter7.AuthenticationInit"?>
+<!-- authentication init have to locate before composition -->
+<?init class="org.zkoss.zk.ui.util.Composition" arg0="/chapter7/layout/template.zul"?>
+
+<zk>
+	<include id="mainInclude" self="@define(content)" src="/chapter7/home.zul"/>
+</zk>
