+++ src/main/webapp/chapter7/layout/banner.zul
@@ -7,7 +7,7 @@
 			<label value="Application Name" sclass="banner-head" />
 		</div>
 		
-		<hbox apply="org.zkoss.essentials.chapter8.LogoutController" 
+		<hbox apply="org.zkoss.essentials.chapter7.LogoutController" 
 			hflex="1" vflex="1" pack="end" align="end" >
 			<label value="${sessionScope.userCredential.name}" if="${not sessionScope.userCredential.anonymous}"/>
 			<label id="logout" value="Logout" if="${not sessionScope.userCredential.anonymous}" sclass="logout"/>
