+++ src/main/webapp/chapter7/layout/template-anonymous.zul
@@ -2,7 +2,7 @@
 	<!-- free to access template, without sidebar  -->
 	<borderlayout hflex="1" vflex="1">
 		<north height="100px" border="none" >
-			<include src="/chapter8/layout/banner.zul"/>
+			<include src="/chapter7/layout/banner.zul"/>
 		</north>
 		<center id="mainContent" autoscroll="true" border="none" self="@insert(content)">
 			<!-- the main content will be insert to here -->
