+++ src/main/webapp/chapter7/layout/template.zul
@@ -2,10 +2,10 @@
 	<!-- create only when the currentUser is not an anonymous  -->
 	<borderlayout hflex="1" vflex="1" if="${not sessionScope.userCredential.anonymous}">
 		<north height="100px" border="none" >
-			<include src="/chapter8/layout/banner.zul"/>
+			<include src="/chapter7/layout/banner.zul"/>
 		</north>
 		<west width="260px" border="none" collapsible="true" splittable="true" minsize="300">
-			<include src="/chapter6/ajaxbased/layout/sidebar.zul"/>
+			<include src="/chapter6/ajaxbased/sidebar.zul"/>
 		</west>
 		<center id="mainContent" autoscroll="true" border="none" self="@insert(content)">
 			<!-- the main content will be insert to here -->
